[refac] chaneg integer type

+++ src/lines.rs
@@ -13,5 +13,5 @@ pub struct Spec {
     pub intense: RVector,
     pub tau: RVector,
     pub stokes: [f64; 3],
-    pub num_rays: i64,
+    pub num_rays: usize,
 }
